REF: Split threads and processes utils into separate modules
